+++ cmake/rtrp.cmake
@@ -3,7 +3,7 @@ include(FetchContent)
 FetchContent_Declare(
     RTRP
     GIT_REPOSITORY https://github.com/SpaghettDev/RTRP.git
-    GIT_TAG        deb04e500252058187b8afa1b1222d9fc5ed3191
+    GIT_TAG        a6f1bb12371db4028b9e7119f4bf44c8ca9c3445
     GIT_PROGRESS TRUE
 )
 message("Fetching RTRP")