+++ src/custom_layers/base/BaseCustomAlertLayer.hpp
@@ -1,4 +1,5 @@
 #pragma once
+#include <limits>
 
 #include <Geode/Bindings.hpp>
 
@@ -10,7 +11,7 @@ public:
 	CCPoint alertSize{};
 	CCMenuItemSpriteExtra* closeBtn{};
 
-	bool createBasics(CCPoint, SEL_MenuHandler, float = 1.f, const ccColor4B& = { 0x00, 0x00, 0x00, 0x4B });
+	bool createBasics(CCPoint, SEL_MenuHandler, float = 1.f, const ccColor4B& = { 0x00, 0x00, 0x00, 0x4B }, int = std::numeric_limits<int>::max());
 	void createTitle(std::string, float = .75f, float = 1.f);
 	CCMenuItemSpriteExtra* createButton(const char*, CCPoint, SEL_MenuHandler, int = -1, float = 1.f, float = 1.2f);
 	virtual void onClose(CCObject*) = 0;